requirements.txt
pyyaml==6.0
tqdm==4.65.0
torch==2.2.1
numpy==1.26.4 
matplotlib==3.8.4
torchvision==0.17.1